adding conftest.py
